added edit mode

+++ src/app/components/task-list-item/task-list-item.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, inject, Input, Output } from '@angular/core';
 import { NgClass } from '@angular/common';
 import { TaskService } from '../../services/task.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-task-list-item',
@@ -15,19 +16,28 @@ export class TaskListItemComponent {
   @Input({ required: true }) priority!: string;
   @Output() emitOnDelete = new EventEmitter<void>();
 
+  private router = inject(Router);
   private taskService = inject(TaskService);
 
   OnDelete(event: Event, itemId: number) {
     event.stopPropagation();
-    this.taskService.RemoveTaskItem(itemId);
+    const isConfirmed: boolean = confirm(
+      'Are you sure you want to delete this task?',
+    );
+
+    if (isConfirmed) {
+      this.taskService.RemoveTaskItem(itemId);
+    } else {
+      return;
+    }
+
     this.emitOnDelete.emit();
   }
 
   OnEdit(event: Event) {
+    this.router.navigate(['/tasks/edit', this.id]);
     event.stopPropagation();
   }
-
-  ViewTask() {
-    console.log('view task');
-  }
+  GoToEditTask(taskId: number) {}
+  ViewTask() {}
 }
